Add Alpaka implementation of ParticleFlow RecHit producer
The PFRecHitProducerAlpaka constructs PFRecHits in the PFRecHitSoA format. Its input are calorimeter hits in the CaloRecHitSoA format and topology information and cut parameters, as supplied by the corresponding ESProducers.
The module is capable of consuming multiple CaloRecHitSoAs, as is required for ECAL, where barrel and endcap hits are supplied separately.